Переход на unsigned G и добавление точности стравнения E и M

+++ headers/adapterGPU.h
@@ -3,11 +3,11 @@
 
 #include "lattice.h"
 
-void latticeConstructorDOSAdapter(long long int *&G, float *&E, float *&M, int size);
+void latticeConstructorDOSAdapter(unsigned long long int *&G, float *&E, float *&M, int size);
 void latticeConstructorAdapter(float *&x, float *&y, float *&mx, float *&my, int size);
-void latticeDestructorAdapter(long long int *&Gmain, float *&Emain, float *&Mmain, 
-    long long int *&Gresult, float *&Eresult, float *&Mresult,  float *&x, float *&y, float *&mx, float *&my);
-void calculateAdapter(long long int *&G, float *&E, float *&M, 
-    float *&x, float *&y, float *&mx, float *&my, int latticeSize, float splitSeed);
+void latticeDestructorAdapter(unsigned long long int *&Gmain, float *&Emain, float *&Mmain, 
+    unsigned long long int *&Gresult, float *&Eresult, float *&Mresult,  float *&x, float *&y, float *&mx, float *&my);
+void calculateAdapter(unsigned long long int *&G, float *&E, int *&M,
+     float *&x, float *&y, float *&mx, float *&my, int latticeSize, float splitSeed);
 
 #endif
